requirements.txt
beautifulsoup4==4.12.3
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
flake8==7.1.1
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
prettytable==3.11.0
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.3
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
soupsieve==2.6
# Editable install with no version control (stock_tracker==0.1.0)
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
